website: Add Nunito and remove unused Search icon.
Nunito is the standard font across our site so it makes sense to update the
changelog with it.

The search feature was never actually implemented so it doesn't work and is
just taking up space.

The entire changelog needs to be redone so these are minimualistic changes but
no major changes are coming to it until we redesign.

Change-Id: I5f6ec049720372508495fe22f3d24263f129c997